refactor: make ACVM rely on blackbox_solver impls
