+++ acvm/Cargo.toml
@@ -21,12 +21,6 @@ stdlib.workspace = true
 brillig_vm = { version = "0.17.0", path = "../brillig_vm", default-features = false }
 blackbox_solver.workspace = true
 
-blake2.workspace = true
-sha2.workspace = true
-sha3.workspace = true
-k256.workspace = true
-p256.workspace = true
-
 indexmap = "1.7.0"
 async-trait = "0.1"
