+++ blackbox_solver/Cargo.toml
@@ -13,12 +13,25 @@ repository.workspace = true
 [dependencies]
 acir_field.workspace = true
 acir.workspace = true
-blake2.workspace = true
-sha2.workspace = true
-sha3.workspace = true
-k256.workspace = true
-p256.workspace = true
 thiserror.workspace = true
+
+blake2 = "0.10.6"
+sha2 = "0.10.6"
+sha3 = "0.10.6"
+k256 = { version = "0.11.0", features = [
+    "ecdsa",
+    "ecdsa-core",
+    "sha256",
+    "digest",
+    "arithmetic",
+] }
+p256 = { version = "0.11.0", features = [
+    "ecdsa",
+    "ecdsa-core",
+    "sha256",
+    "digest",
+    "arithmetic",
+] }
 
 [features]
 default = ["bn254"]
